Handle Date/Time as ISO-8601
To properly handle date and time in Javascript this handles the SQL date and time return values appropriately as ISO-8601 formatted values.  Since there is no way to know the timezone used in SQL it is up to the developer to add an appropriate timezone offset.  If someone can conceive of a way to provide the compiler with timezone information for return values in one or more columns of the multiple record sets on return I'd be interested in seeing it.

Potential Breaking Change: Since the original code returned the date/time using .ToString() then it would be in a format that would have been presentable to the user in the same timezone/culture.  Results from the query will not be in ISO-8601 format which means that the Javascript will be responsible for formatting the output appropriately.